+++ src/banco_de_dados/dao/postgresql/EspecialidadeDAOPostgresql.java
@@ -19,19 +19,49 @@ public class EspecialidadeDAOPostgresql extends ConectorDAOPostgresql implements
     
     
     @Override
-    public Especialidade criar(int codigo, int indice, String nome) throws BancoDeDadosException, SQLException {
-        
-        Especialidade especialidade = new Especialidade(codigo, indice, nome);
+    public boolean existeEspecialidade(int codigo) throws BancoDeDadosException, SQLException {
+    
         Connection conexao = this.fabricaDeConexoes.getConexao();
         
-        String sql = "INSERT INTO " + NOME_COMPLETO + " VALUES (?, ?, ?)";
+        String sql = "SELECT COUNT(1) FROM " + NOME_COMPLETO + 
+                " WHERE codigo =" + codigo;
+        
+        try(ResultSet resultSet = conexao.createStatement().executeQuery(sql)) {
+            if(resultSet.next()) {
+                int quantidade = resultSet.getInt(1);
+                
+                return quantidade != 0;
+            }
+        } catch(SQLException sqle) {
+            throw new BancoDeDadosException("Não foi possível verificar a exitência da especialidade no banco de dados", sqle);
+        } finally {
+            conexao.close();
+        }
+        
+        return false;
+    }
+    
+    @Override
+    public Especialidade criar(int indice, String nome) throws BancoDeDadosException, SQLException {
+        
+        Especialidade especialidade = new Especialidade(indice, nome);
+        Connection conexao = this.fabricaDeConexoes.getConexao();
+        
+        String sql = "INSERT INTO " + NOME_COMPLETO + " VALUES (DEFAULT, ?, ?)"
+                + " RETURNING codigo";
         
         try(PreparedStatement statement = conexao.prepareStatement(sql)) {
-            statement.setInt(1, codigo);
-            statement.setInt(2, indice);
-            statement.setString(3, nome);
+            statement.setInt(1, indice);
+            statement.setString(2, nome);
 
             statement.execute();
+            
+            ResultSet resultSet = statement.getResultSet();
+            
+            if(resultSet.next()) {
+                especialidade.setCodigo(resultSet.getInt(1));
+            }
+            
             statement.close();
         } catch(SQLException sqle) {
             throw new BancoDeDadosException("Não foi possível inserir a especialidade no banco de dados", sqle);
@@ -66,6 +96,10 @@ public class EspecialidadeDAOPostgresql extends ConectorDAOPostgresql implements
     @Override
     public void remover(int codigo) throws BancoDeDadosException, SQLException {
         
+        if(!this.existeEspecialidade(codigo)) {
+            throw new BancoDeDadosException("Não existe nenhuma especialidade com este código");
+        }
+        
         Connection conexao = this.fabricaDeConexoes.getConexao();
         
         String sql = "DELETE FROM " + NOME_COMPLETO + " WHERE codigo = ?";
@@ -92,11 +126,7 @@ public class EspecialidadeDAOPostgresql extends ConectorDAOPostgresql implements
         String sql = "SELECT * FROM " + NOME_COMPLETO + "  WHERE codigo = " + codigo;
         try(ResultSet resultSet = conexao.createStatement().executeQuery(sql)) {
             if(resultSet.next()) {
-                int codigoEncontrado = resultSet.getInt("codigo");
-                int indice = resultSet.getInt("indice");
-                String nome = resultSet.getString("nome");
-
-                especialidade = new Especialidade(codigoEncontrado, indice, nome);
+                especialidade = this.criarEspecialidadeAPartirDe(resultSet);
             }
         } catch(SQLException sqle) {
             throw new BancoDeDadosException("Não foi possível encontrar esta especialidade no banco de dados", sqle);
@@ -117,11 +147,8 @@ public class EspecialidadeDAOPostgresql extends ConectorDAOPostgresql implements
         String sql = "SELECT * FROM " + NOME_COMPLETO + "  WHERE nome LIKE '%" + nome + "%'";
         try(ResultSet resultSet = conexao.createStatement().executeQuery(sql)) {
             while(resultSet.next()) {
-                int codigo = resultSet.getInt("codigo");
-                int indice = resultSet.getInt("indice");
-                String nomeEncontrado = resultSet.getString("nome");
-
-                especialidadesEncontradas.add(new Especialidade(codigo, indice, nomeEncontrado));
+                Especialidade especialidade = this.criarEspecialidadeAPartirDe(resultSet);
+                especialidadesEncontradas.add(especialidade);
             }
         } catch(SQLException sqle) {
             throw new BancoDeDadosException("Erro ao buscar especialidade(s) no banco de dados", sqle);
@@ -130,5 +157,14 @@ public class EspecialidadeDAOPostgresql extends ConectorDAOPostgresql implements
         }
         
         return especialidadesEncontradas;
+    }
+    
+    
+    private Especialidade criarEspecialidadeAPartirDe(ResultSet resultSet) throws SQLException {
+         int codigo = resultSet.getInt("codigo");
+         int indice = resultSet.getInt("indice");
+         String nome = resultSet.getString("nome");
+         
+         return new Especialidade(codigo, indice, nome);
     }
 }
